Rename wrapper.py as app.py
